Added commit

+++ homework2/task2.hpp
@@ -1,7 +1,8 @@
 #ifndef TASK2_HPP
 #define TASK2_HPP
+#include <mutex>
 
-void drawRowProgress(int, int);
+void drawRowProgress(size_t, int);
 void task2(int);
 
 #endif
